JUnit 5 Migration (#1163)
* Replace JUnit 4 dependency with JUnit 5

* Replace JUnit 4 imports with JUnit 5 equivalents

* Use JUnit 5 fail

* Test exceptions with JUnit 5

* Replace JUnit 3 imports with JUnit 5 equivalents

* Remove expected attribute in @Test

expected attribute is no longer supported in JUnit 5.

* Fix broken JUnit imports

* Change failure message positions

Unlike JUnit 4, JUnit 5 assert methods require the message
to be positioned after the condition

* Make more direct assertion calls

* Fix broken assertion call

* Replace Hamcrest assert calls with JUnit 5

* Use Java builtin constants / methods for test IO

* Replace JUnit 4 annotations with JUnit 5 equivalents

@Before -> @BeforeEach
@BeforeClass -> @BeforeAll
@AfterClass -> @AfterAll

* Use JUnit 5 ParameterizedTest

* Replace JUnit4 category with JUnit 5 tag

* Add missing arguments to test

* Replace JUnit Toolbox with JUnit Platform

* Remove unused JUnit 4 imports

* Restore Hamcrest imports

Hamcrest must be explicitly be declared as a dependency
to be used with JUnit 5

* Replace JUnit 4 annotation with JUnit 5

* Avoid false negative in test

Conversion to primitive int was removed while
migrating to JUnit 5, and this commit restores it

* Adjust Surefire / Failsafe configurations

* Temporaily include JUnit Vintage (JUnit 4)

* Rewrite parallelized test code

- Remove JUnit 3 dependencies
- Manually construct threads / streams to execute concurrent code

* Remove JUnit Vintage (JUnit 4) dependency

This reverts commit 36c3169f037fa7ffe76e9e29924269cc0547356e.

* Cleanup JUnit 5 Assertion calls

- Position expected value before actual
- Avoid using wrapper -> primitive conversion (built in JUnit 5)

* Use @RepeatedTest annotation

* Update testing plugin versions

* Properly set up JUnit 5 Suites

- Add @Suite tags to test suite classes
- Properly tell Surefire / Failsafe how to find the suites

* Run tests in parallel

* Remove unused JUnit imports

* Update dependency description in README

JGrapht no longer uses JUnit 4 for testing

* Specify which package to include in test suite

* Properly setup JUnit 5 test suites in pom

* Mark argument provider method public

JUnit 5 engine seems quite picky with access
modifiers when combined with Java modules.

The simplest way around it is marking it public.

* Manually export package to unnamed module

Because org.jgrapht.interfaces package is not exported by
org.jgrapht.core module, this creates visibility error in AStarAdmissibleHeuristicTest

Ref: https://github.com/junit-team/junit5/issues/2147

* Reset BergeGraphInspector before every test run

* Restore test filtering in Surefire / Failsafe

* Disable failIfNoSpecifiedTests in Surefire

Ref: https://github.com/jgrapht/jgrapht/pull/1163#discussion_r1432320707

* Manually construct threads in AsSynchronizedGraphTest

Ref: https://github.com/jgrapht/jgrapht/pull/1163#issuecomment-1865715293

* Force tests in AsSynchronizedGraph to run on same thread

* Replace AtomicInteger with Semaphore

* Use correct annotation for PerformanceTestSuite

+++ jgrapht-core/src/test/java/org/jgrapht/alg/color/ColorRefinementAlgorithmTest.java
@@ -19,12 +19,12 @@ package org.jgrapht.alg.color;
 
 import org.jgrapht.*;
 import org.jgrapht.graph.*;
-import org.junit.*;
+import org.junit.jupiter.api.*;
 
 import java.util.*;
 
-import static org.junit.Assert.assertEquals;
-import static org.junit.Assert.assertNotEquals;
+import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertNotEquals;
 
 /**
  * Tests for the color-refinement algorithm.
@@ -56,13 +56,13 @@ public class ColorRefinementAlgorithmTest
 
         // symmetric pairs around 3 should have the same color and different colors otherwise
 
-        assertEquals(colors.get(1).intValue(), colors.get(5).intValue());
-        assertNotEquals(colors.get(1).intValue(), colors.get(7).intValue());
-        assertEquals(colors.get(2).intValue(), colors.get(4).intValue());
-        assertEquals(colors.get(7).intValue(), colors.get(8).intValue());
-        assertNotEquals(colors.get(1).intValue(), colors.get(2).intValue());
-        assertNotEquals(colors.get(2).intValue(), colors.get(3).intValue());
-        assertNotEquals(colors.get(3).intValue(), colors.get(6).intValue());
+        assertEquals(colors.get(1), colors.get(5));
+        assertNotEquals(colors.get(1), colors.get(7));
+        assertEquals(colors.get(2), colors.get(4));
+        assertEquals(colors.get(7), colors.get(8));
+        assertNotEquals(colors.get(1), colors.get(2));
+        assertNotEquals(colors.get(2), colors.get(3));
+        assertNotEquals(colors.get(3), colors.get(6));
     }
 
     @Test
@@ -88,11 +88,11 @@ public class ColorRefinementAlgorithmTest
 
         // all vertices should have the same color
 
-        assertEquals(colors.get(1).intValue(), colors.get(2).intValue());
-        assertEquals(colors.get(1).intValue(), colors.get(3).intValue());
-        assertEquals(colors.get(1).intValue(), colors.get(4).intValue());
-        assertEquals(colors.get(1).intValue(), colors.get(5).intValue());
-        assertEquals(colors.get(1).intValue(), colors.get(6).intValue());
+        assertEquals(colors.get(1), colors.get(2));
+        assertEquals(colors.get(1), colors.get(3));
+        assertEquals(colors.get(1), colors.get(4));
+        assertEquals(colors.get(1), colors.get(5));
+        assertEquals(colors.get(1), colors.get(6));
     }
 
     @Test
@@ -136,11 +136,11 @@ public class ColorRefinementAlgorithmTest
         for (int i = 1; i < 11; i++) {
             for (int j = i + 1; j <= 11; j++) {
                 if (i != 9 || j != 10) {
-                    assertNotEquals(colors.get(i).intValue(), colors.get(j).intValue());
+                    assertNotEquals(colors.get(i), colors.get(j));
                 }
             }
         }
-        assertEquals(colors.get(9).intValue(), colors.get(10).intValue());
+        assertEquals(colors.get(9), colors.get(10));
     }
 
     @Test
@@ -177,12 +177,12 @@ public class ColorRefinementAlgorithmTest
         for (int i = 1; i < 9; i++) {
             for (int j = i + 1; j < 9; j++) {
                 if ((i == 1 && j == 3) || (i == 5 && j == 7)) {
-                    assertEquals(colors.get(i).intValue(), colors.get(j).intValue());
+                    assertEquals(colors.get(i), colors.get(j));
                 } else {
-                    assertNotEquals(colors.get(i).intValue(), colors.get(j).intValue());
+                    assertNotEquals(colors.get(i), colors.get(j));
                 }
             }
         }
-        assertEquals(colors.get(1).intValue(), colors.get(3).intValue());
+        assertEquals(colors.get(1), colors.get(3));
     }
 }
