+++ jgrapht-core/src/test/java/org/jgrapht/alg/color/LargestDegreeFirstColoringTest.java
@@ -21,11 +21,11 @@ import org.jgrapht.*;
 import org.jgrapht.alg.interfaces.*;
 import org.jgrapht.alg.interfaces.VertexColoringAlgorithm.*;
 import org.jgrapht.graph.*;
-import org.junit.*;
+import org.junit.jupiter.api.*;
 
 import java.util.*;
 
-import static org.junit.Assert.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertEquals;
 
 /**
  * Coloring tests
@@ -72,11 +72,11 @@ public class LargestDegreeFirstColoringTest
         Coloring<Integer> coloring = new LargestDegreeFirstColoring<>(g).getColoring();
         assertEquals(3, coloring.getNumberColors());
         Map<Integer, Integer> colors = coloring.getColors();
-        assertEquals(0, colors.get(1).intValue());
-        assertEquals(2, colors.get(2).intValue());
-        assertEquals(1, colors.get(3).intValue());
-        assertEquals(2, colors.get(4).intValue());
-        assertEquals(2, colors.get(5).intValue());
+        assertEquals(0, colors.get(1));
+        assertEquals(2, colors.get(2));
+        assertEquals(1, colors.get(3));
+        assertEquals(2, colors.get(4));
+        assertEquals(2, colors.get(5));
     }
 
     @Test
@@ -98,12 +98,12 @@ public class LargestDegreeFirstColoringTest
         Coloring<Integer> coloring = new LargestDegreeFirstColoring<>(g).getColoring();
         assertEquals(3, coloring.getNumberColors());
         Map<Integer, Integer> colors = coloring.getColors();
-        assertEquals(1, colors.get(1).intValue());
-        assertEquals(2, colors.get(2).intValue());
-        assertEquals(0, colors.get(3).intValue());
-        assertEquals(2, colors.get(4).intValue());
-        assertEquals(2, colors.get(5).intValue());
-        assertEquals(1, colors.get(6).intValue());
+        assertEquals(1, colors.get(1));
+        assertEquals(2, colors.get(2));
+        assertEquals(0, colors.get(3));
+        assertEquals(2, colors.get(4));
+        assertEquals(2, colors.get(5));
+        assertEquals(1, colors.get(6));
     }
 
     @Test
@@ -121,12 +121,12 @@ public class LargestDegreeFirstColoringTest
         Coloring<Integer> coloring = new LargestDegreeFirstColoring<>(g).getColoring();
         assertEquals(3, coloring.getNumberColors());
         Map<Integer, Integer> colors = coloring.getColors();
-        assertEquals(0, colors.get(1).intValue());
-        assertEquals(0, colors.get(2).intValue());
-        assertEquals(1, colors.get(3).intValue());
-        assertEquals(2, colors.get(4).intValue());
-        assertEquals(0, colors.get(5).intValue());
-        assertEquals(1, colors.get(6).intValue());
+        assertEquals(0, colors.get(1));
+        assertEquals(0, colors.get(2));
+        assertEquals(1, colors.get(3));
+        assertEquals(2, colors.get(4));
+        assertEquals(0, colors.get(5));
+        assertEquals(1, colors.get(6));
     }
 
     @Test
@@ -143,12 +143,12 @@ public class LargestDegreeFirstColoringTest
         Coloring<Integer> coloring = new LargestDegreeFirstColoring<>(g).getColoring();
         assertEquals(3, coloring.getNumberColors());
         Map<Integer, Integer> colors = coloring.getColors();
-        assertEquals(0, colors.get(1).intValue());
-        assertEquals(0, colors.get(2).intValue());
-        assertEquals(1, colors.get(3).intValue());
-        assertEquals(1, colors.get(4).intValue());
-        assertEquals(0, colors.get(5).intValue());
-        assertEquals(2, colors.get(6).intValue());
+        assertEquals(0, colors.get(1));
+        assertEquals(0, colors.get(2));
+        assertEquals(1, colors.get(3));
+        assertEquals(1, colors.get(4));
+        assertEquals(0, colors.get(5));
+        assertEquals(2, colors.get(6));
     }
 
 }
