+++ jgrapht-core/src/test/java/org/jgrapht/alg/flow/MaximumFlowAlgorithmTest.java
@@ -20,12 +20,12 @@ package org.jgrapht.alg.flow;
 import org.jgrapht.*;
 import org.jgrapht.alg.interfaces.*;
 import org.jgrapht.graph.*;
-import org.junit.*;
+import org.junit.jupiter.api.*;
 
 import java.util.*;
 
-import static org.junit.Assert.assertEquals;
-import static org.junit.Assert.assertTrue;
+import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertTrue;
 
 /**
  * @author Joris Kinable
@@ -113,7 +113,7 @@ public abstract class MaximumFlowAlgorithmTest
         Double flowValue = maxFlow.getValue();
         Map<DefaultWeightedEdge, Double> flow = maxFlow.getFlowMap();
 
-        assertEquals(expectedResult, flowValue.intValue());
+        assertEquals(expectedResult, flowValue);
 
         // Verify that every edge is contained in the flow map
         for (DefaultWeightedEdge e : graph.edgeSet())
