+++ jgrapht-core/src/test/java/org/jgrapht/alg/shortestpath/TreeSingleSourcePathsTest.java
@@ -20,12 +20,12 @@ package org.jgrapht.alg.shortestpath;
 import org.jgrapht.*;
 import org.jgrapht.alg.util.*;
 import org.jgrapht.graph.*;
-import org.junit.*;
+import org.junit.jupiter.api.*;
 
 import java.util.*;
 
-import static org.junit.Assert.assertEquals;
-import static org.junit.Assert.assertTrue;
+import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertTrue;
 
 /**
  * @author Dimitrios Michail
@@ -66,7 +66,7 @@ public class TreeSingleSourcePathsTest
         TreeSingleSourcePathsImpl<Integer, DefaultWeightedEdge> t1 =
             new TreeSingleSourcePathsImpl<>(g, 1, map);
 
-        assertEquals(1, t1.getSourceVertex().intValue());
+        assertEquals(1, t1.getSourceVertex());
         assertEquals(0d, t1.getWeight(1), 1e-9);
         assertTrue(t1.getPath(1).getEdgeList().isEmpty());
         assertEquals(Arrays.asList(g.getEdgeSource(e12_1)), t1.getPath(1).getVertexList());
